+++ metasvc-model/src/main/java/de/digitalcollections/model/exception/problem/ProblemHint.java
@@ -19,4 +19,9 @@ public enum ProblemHint {
   public String getDescription() {
     return description;
   }
+
+  @Override
+  public String toString() {
+    return name() + " (" + description + ")";
+  }
 }